move createzip fct to Album controller
